+++ packages/dashboard-core-plugins/src/ChartWidgetPlugin.tsx
@@ -1,14 +1,17 @@
 import { useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
 import { useApi } from '@deephaven/jsapi-bootstrap';
 import { Chart, type ChartModel, ChartModelFactory } from '@deephaven/chart';
 import type { dh } from '@deephaven/jsapi-types';
 import { type WidgetComponentProps } from '@deephaven/plugin';
+import { getSettings, type RootState } from '@deephaven/redux';
 
 export function ChartWidgetPlugin(
   props: WidgetComponentProps<dh.plot.Figure>
 ): JSX.Element | null {
   const dh = useApi();
   const [model, setModel] = useState<ChartModel>();
+  const settings = useSelector(getSettings<RootState>);
 
   const { fetch } = props;
 
@@ -30,7 +33,7 @@ export function ChartWidgetPlugin(
     };
   }, [dh, fetch]);
 
-  return model ? <Chart model={model} /> : null;
+  return model ? <Chart model={model} settings={settings} /> : null;
 }
 
 export default ChartWidgetPlugin;