+++ packages/iris-grid/src/sidebar/index.ts
@@ -4,7 +4,10 @@ import AggregationEdit from './aggregations/AggregationEdit';
 import AggregationUtils from './aggregations/AggregationUtils';
 import ChartBuilder from './ChartBuilder';
 import CustomColumnBuilder from './CustomColumnBuilder';
-import OptionType from './OptionType';
+import OptionType, {
+  type OptionItemKey,
+  type PluginOptionKey,
+} from './OptionType';
 import RollupRows from './RollupRows';
 import TableCsvExporter from './TableCsvExporter';
 import TableSaver from './TableSaver';
@@ -12,6 +15,10 @@ import VisibilityOrderingBuilder from './visibility-ordering-builder/VisibilityO
 import { type FormattingRule } from './conditional-formatting/ConditionalFormattingUtils';
 import AdvancedSettingsType from './AdvancedSettingsType';
 import DownloadServiceWorkerUtils from './DownloadServiceWorkerUtils';
+import {
+  type IrisGridTableOptionsWidgetProps,
+  type TableOptionsTransform,
+} from './IrisGridTableOptionsWidgetProps';
 
 export {
   AdvancedSettings,
@@ -21,12 +28,16 @@ export {
   ChartBuilder,
   CustomColumnBuilder,
   OptionType,
+  type OptionItemKey,
+  type PluginOptionKey,
   RollupRows,
   TableCsvExporter,
   TableSaver,
   VisibilityOrderingBuilder,
   AdvancedSettingsType,
   DownloadServiceWorkerUtils,
+  type IrisGridTableOptionsWidgetProps,
+  type TableOptionsTransform,
 };
 
 export type { FormattingRule as SidebarFormattingRule };